Issue #4313: Split and Organize Checkstyle inputs by Test for checks in annotation package: SuppressWarnings
